+++ src/anemoi/inference/outputs/netcdf.py
@@ -32,19 +32,7 @@ LOCK = threading.RLock()
 @output_registry.register("netcdf")
 @main_argument("path")
 class NetCDFOutput(Output):
-    """NetCDF output class.
-
-    Parameters
-    ----------
-    context : dict
-        The context dictionary.
-    path : str
-        The path to save the NetCDF file.
-    output_frequency : int, optional
-        The frequency of output, by default None.
-    write_initial_state : bool, optional
-        Whether to write the initial state, by default None.
-    """
+    """NetCDF output class."""
 
     def __init__(
         self,
@@ -53,19 +41,39 @@ class NetCDFOutput(Output):
         variables: Optional[List[str]] = None,
         output_frequency: Optional[int] = None,
         write_initial_state: Optional[bool] = None,
+        float_size: str = "f4",
+        missing_value: Optional[float] = np.nan,
     ) -> None:
-        super().__init__(
-            context,
-            variables=variables,
-            output_frequency=output_frequency,
-            write_initial_state=write_initial_state,
-        )
+        """Initialize the NetCDF output object.
+
+        Parameters
+        ----------
+        context : dict
+            The context dictionary.
+        path : str
+            The path to save the NetCDF file.
+        output_frequency : int, optional
+            The frequency of output, by default None.
+        write_initial_state : bool, optional
+            Whether to write the initial state, by default None.
+        float_size : str, optional
+            The size of the float, by default "f4".
+        missing_value : float, optional
+            The missing value, by default np.nan.
+        """
+
+        super().__init__(context, output_frequency=output_frequency, write_initial_state=write_initial_state)
 
         from netCDF4 import Dataset
 
         self.path = path
         self.ncfile: Optional[Dataset] = None
-        self.float_size = "f4"
+        self.float_size = float_size
+        self.missing_value = missing_value
+        if self.write_step_zero:
+            self.extra_time = 1
+        else:
+            self.extra_time = 0
 
     def __repr__(self) -> str:
         """Return a string representation of the NetCDFOutput object."""
@@ -102,6 +110,8 @@ class NetCDFOutput(Output):
             lead_time := getattr(self.context, "lead_time", None)
         ):
             time = lead_time // time_step
+            time += self.extra_time
+
         if reference_date := getattr(self.context, "reference_date", None):
             self.reference_date = reference_date
 
@@ -130,7 +140,6 @@ class NetCDFOutput(Output):
         self.longitude_var[:] = longitudes
 
         self.vars = {}
-        self.ensure_variables(state)
 
         self.n = 0
 
@@ -142,6 +151,7 @@ class NetCDFOutput(Output):
         state : State
             The state dictionary.
         """
+
         values = len(state["latitudes"])
 
         compression = {}  # dict(zlib=False, complevel=0)
@@ -160,13 +170,19 @@ class NetCDFOutput(Output):
                 chunksizes = tuple(int(np.ceil(x / 2)) for x in chunksizes)
 
             with LOCK:
+                missing_value = self.missing_value
+
                 self.vars[name] = self.ncfile.createVariable(
                     name,
                     self.float_size,
                     ("time", "values"),
                     chunksizes=chunksizes,
+                    fill_value=missing_value,
                     **compression,
                 )
+
+                self.vars[name].fill_value = missing_value
+                self.vars[name].missing_value = missing_value
 
     def write_step(self, state: State) -> None:
         """Write the state.
@@ -188,6 +204,7 @@ class NetCDFOutput(Output):
                 continue
 
             with LOCK:
+                LOG.info(f"🚧🚧🚧🚧🚧🚧 XXXXXX {name}, {self.n}, {value.shape}")
                 self.vars[name][self.n] = value
 
         self.n += 1