+++ src/anemoi/inference/outputs/tee.py
@@ -55,12 +55,7 @@ class TeeOutput(ForwardOutput):
         **kwargs : Any
             Additional keyword arguments.
         """
-        super().__init__(
-            context,
-            variables=variables,
-            output_frequency=output_frequency,
-            write_initial_state=write_initial_state,
-        )
+        super().__init__(context, None, output_frequency=output_frequency, write_initial_state=write_initial_state)
 
         if outputs is None:
             outputs = args